* More refactoring
