+++ MatchingMakingMonitor/Models/Replay/Vehicle.cs
@@ -1,0 +1,11 @@
+// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.Replay
+{
+	public class Vehicle
+	{
+		public long id { get; set; }
+		public int relation { get; set; }
+		public string name { get; set; }
+		public long shipId { get; set; }
+	}
+}
