+++ MatchingMakingMonitor/Models/ResponseTypes/Aircraft.cs
@@ -1,0 +1,9 @@
+// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class Aircraft
+	{
+		public int max_frags_battle { get; set; }
+		public int frags { get; set; }
+	}
+}
