+++ MatchingMakingMonitor/Models/ResponseTypes/Data.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class Data
+	{
+		public List<Ship> Ships { get; set; }
+	}
+}
