+++ MatchingMakingMonitor/Models/ResponseTypes/Meta.cs
@@ -1,0 +1,9 @@
+// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class Meta
+	{
+		public int count { get; set; }
+		public object hidden { get; set; }
+	}
+}
