+++ MatchingMakingMonitor/Models/ResponseTypes/PlayerStats.cs
@@ -1,0 +1,10 @@
+﻿// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class PlayerStats
+	{
+		public string status { get; set; }
+		public Meta meta { get; set; }
+		public Data data { get; set; }
+	}
+}
