+++ MatchingMakingMonitor/Models/ResponseTypes/Ramming.cs
@@ -1,0 +1,9 @@
+// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class Ramming
+	{
+		public int max_frags_battle { get; set; }
+		public int frags { get; set; }
+	}
+}
