+++ MatchingMakingMonitor/Models/ResponseTypes/SecondBattery.cs
@@ -1,0 +1,11 @@
+// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class SecondBattery
+	{
+		public int max_frags_battle { get; set; }
+		public int frags { get; set; }
+		public int hits { get; set; }
+		public int shots { get; set; }
+	}
+}
