+++ MatchingMakingMonitor/Models/ResponseTypes/Torpedoes.cs
@@ -1,0 +1,11 @@
+// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class Torpedoes
+	{
+		public int max_frags_battle { get; set; }
+		public int frags { get; set; }
+		public int hits { get; set; }
+		public int shots { get; set; }
+	}
+}
