+++ MatchingMakingMonitor/Models/ResponseTypes/WargamingPlayer.cs
@@ -1,0 +1,10 @@
+// ReSharper disable InconsistentNaming
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class WargamingPlayer
+	{
+		public string nickname { get; set; }
+		public long account_id { get; set; }
+		public string Region { get; set; }
+	}
+}
