+++ MatchingMakingMonitor/Models/ResponseTypes/WargamingSearch.cs
@@ -1,0 +1,12 @@
+﻿// ReSharper disable InconsistentNaming
+using System.Collections.Generic;
+
+namespace MatchMakingMonitor.Models.ResponseTypes
+{
+	public class WargamingSearch
+	{
+		public string status { get; set; }
+		public Dictionary<string, int> meta { get; set; }
+		public List<WargamingPlayer> data { get; set; }
+	}
+}
