+++ MatchingMakingMonitor/Models/ShipInfo.cs
@@ -6,5 +6,5 @@
         public string ShipName { get; set; }
         public string ShipType { get; set; }
         public string Tier { get; set; }
-    } //end class
-} //end namespace
+    }
+}
