+++ MatchingMakingMonitor/Services/ApiService.cs
@@ -7,6 +7,8 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
+using MatchMakingMonitor.Models.Replay;
+using MatchMakingMonitor.Models.ResponseTypes;
 
 namespace MatchMakingMonitor.Services
 {
