+++ MatchingMakingMonitor/Services/StatsService.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
 using System.Threading.Tasks;
+using MatchMakingMonitor.Models.Replay;
 
 namespace MatchMakingMonitor.Services
 {
